Fix calculation logic for basal chart. Move basal chart provider protocol to separate file.
