+++ xDrip/Model/Entries/Workers/InsulinEntriesWorker.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import AKUtils
 
 final class InsulinEntriesWorker: AbstractEntriesWorker {
     @discardableResult static func addBolusEntry(amount: Double, date: Date) -> InsulinEntry {
@@ -33,6 +34,102 @@ final class InsulinEntriesWorker: AbstractEntriesWorker {
     
     static func fetchAllInsulinEntries() -> [InsulinEntry] {
         return super.fetchAllEntries(type: InsulinEntry.self)
+    }
+    
+    static func getBasalValueForDate(date: Date) -> Double {
+        let minimumDate = Date() - .secondsPerDay * 3.0
+        guard date >= minimumDate else { return 0.0 }
+        let all = InsulinEntriesWorker.fetchAllBasalEntries().filter({ $0.date >=? minimumDate })
+        
+        guard !all.isEmpty else { return 0.0 }
+        
+        var lastValue = all[0].amount
+        var prevEntry: InsulinEntry?
+        for (index, entry) in all.enumerated() {
+            if entry.date >? date { break }
+            if index != 0 {
+                calculateValue(prevEntry: prevEntry, toDate: entry.date, lastValue: &lastValue)
+            }
+            prevEntry = entry
+        }
+        calculateValue(prevEntry: prevEntry, toDate: date, lastValue: &lastValue)
+        
+        return lastValue
+    }
+    
+    private static func calculateValue(prevEntry: InsulinEntry?, toDate: Date?, lastValue: inout Double) {
+        guard let prevEntry = prevEntry, let prevDate = prevEntry.date else { return }
+        let basalRates = User.current.settings.sortedBasalRates
+        guard !basalRates.isEmpty else { return }
+        
+        if basalRates.count == 1 {
+            let unitsPerHour = Double(basalRates[0].units)
+            let date = toDate ?? Date()
+            let diff = unitsPerHour * date.timeIntervalSince(prevDate).hours
+            
+            if lastValue - diff >= 0.0 {
+                lastValue -= diff
+            } else {
+                lastValue = 0.0
+            }
+        } else {
+            var lastDate = prevDate
+            let startOfDay = Calendar.current.startOfDay(for: prevDate)
+            let prevEntryStartTime = prevDate.timeIntervalSince(startOfDay)
+            let currentEntryStartTime = (toDate ?? Date()).timeIntervalSince(startOfDay)
+            
+            var rates = calculateRates(startTime: prevEntryStartTime, endTime: currentEntryStartTime)
+            var unitsPerHour = Double(rates[0].units)
+            rates.removeFirst()
+            
+            func calcValue(startTime: TimeInterval) {
+                let time = startOfDay.addingTimeInterval(startTime)
+                let interval = time.timeIntervalSince(lastDate)
+                let diff = unitsPerHour * interval.hours
+                
+                if lastValue - diff >= 0.0 {
+                    lastValue -= diff
+                } else {
+                    lastValue = 0.0
+                }
+            }
+            
+            for index in 0 ..< rates.endIndex {
+                calcValue(startTime: rates[index].startTime)
+            }
+            calcValue(startTime: currentEntryStartTime)
+        }
+    }
+    
+    private static func calculateRates(startTime: TimeInterval, endTime: TimeInterval) -> [BasalRate] {
+        let basalRates = User.current.settings.sortedBasalRates
+        var rates = basalRates
+        for basalRate in basalRates {
+            rates.append(contentsOf:
+                [
+                    BasalRate(startTime: basalRate.startTime - .secondsPerDay * 3.0, units: basalRate.units),
+                    BasalRate(startTime: basalRate.startTime - .secondsPerDay * 2.0, units: basalRate.units),
+                    BasalRate(startTime: basalRate.startTime - .secondsPerDay, units: basalRate.units),
+                    BasalRate(startTime: basalRate.startTime + .secondsPerDay, units: basalRate.units),
+                    BasalRate(startTime: basalRate.startTime + .secondsPerDay * 2.0, units: basalRate.units),
+                    BasalRate(startTime: basalRate.startTime + .secondsPerDay * 3.0, units: basalRate.units)
+                ]
+            )
+        }
+        rates.sort(by: { $0.startTime < $1.startTime })
+        
+        var minIndex = 0
+        var maxIndex = rates.endIndex - 1
+        for index in 0 ..< rates.endIndex - 1 {
+            if rates[index].startTime <= startTime && rates[index + 1].startTime > startTime {
+                minIndex = index
+            }
+            if rates[index].startTime <= endTime, rates[index + 1].startTime > endTime {
+                maxIndex = index
+            }
+        }
+        
+        return Array(rates[minIndex ... maxIndex])
     }
 }
